yaml: support yaml aliases and anchors
Allows the Suricata configuration file to use YAML anchors
and aliases. An anchor (&) allows a YAML section to be given a
name which can be referenced later on with an alias (*).

my-value: &my-value-anchor "the value"
an-alias: *my-value-anchor

The above will give the field "an-alias" the same value as
"my-value". This is not limited scalar values, but can be used
on lists and maps as well.

Merge keys are also supported which allow you to merge the
contents of one map into another.

first-map: &firstmap
  field1: one
  field2: two

second-map:
  <<: *firstmap
  field3: three
  field4: four

will result in "second-map" being a map of the values in first-map
and second-map (first-map is merged into second-map).